+++ FoodVision/Service/DummyFoodProcessor.swift
@@ -49,32 +49,32 @@ class DummyFoodProcessor {
         let meal = MealDataPoint(
             image: image,
             foodName: "Food",
-            calories: Int.random(in: 50...600), // Random calories between 50 and 600
+            calories: Double.random(in: 50...600), // Random calories between 50 and 600
             transFat: Double.random(in: 0...5), // Random trans fat between 0 and 5 grams
             saturatedFat: Double.random(in: 0...20), // Random saturated fat between 0 and 20 grams
             totalFat: Double.random(in: 0...50), // Random total fat between 0 and 50 grams
             protein: Double.random(in: 0...50), // Random protein between 0 and 50 grams
             sugar: Double.random(in: 0...50), // Random sugar between 0 and 50 grams
             cholesterol: Double.random(in: 0...300), // Random cholesterol between 0 and 300 mg
-            sodium: Int.random(in: 0...2000), // Random sodium between 0 and 2000 mg
-            calcium: Int.random(in: 0...1000), // Random calcium between 0 and 1000 mg
-            iodine: Int.random(in: 0...150), // Random iodine between 0 and 150 mcg
-            iron: Int.random(in: 0...30), // Random iron between 0 and 30 mg
-            magnesium: Int.random(in: 0...400), // Random magnesium between 0 and 400 mg
-            potassium: Int.random(in: 0...3000), // Random potassium between 0 and 3000 mg
-            zinc: Int.random(in: 0...40), // Random zinc between 0 and 40 mg
-            vitaminA: Int.random(in: 0...10000), // Random vitamin A between 0 and 10000 IU
-            vitaminC: Int.random(in: 0...2000), // Random vitamin C between 0 and 2000 mg
-            vitaminD: Int.random(in: 0...1000), // Random vitamin D between 0 and 1000 IU
-            vitaminE: Int.random(in: 0...30), // Random vitamin E between 0 and 30 mg
-            vitaminK: Int.random(in: 0...300), // Random vitamin K between 0 and 300 mcg
+            sodium: Double.random(in: 0...2000), // Random sodium between 0 and 2000 mg
+            calcium: Double.random(in: 0...1000), // Random calcium between 0 and 1000 mg
+            iodine: Double.random(in: 0...150), // Random iodine between 0 and 150 mcg
+            iron: Double.random(in: 0...30), // Random iron between 0 and 30 mg
+            magnesium: Double.random(in: 0...400), // Random magnesium between 0 and 400 mg
+            potassium: Double.random(in: 0...3000), // Random potassium between 0 and 3000 mg
+            zinc: Double.random(in: 0...40), // Random zinc between 0 and 40 mg
+            vitaminA: Double.random(in: 0...10000), // Random vitamin A between 0 and 10000 IU
+            vitaminC: Double.random(in: 0...2000), // Random vitamin C between 0 and 2000 mg
+            vitaminD: Double.random(in: 0...1000), // Random vitamin D between 0 and 1000 IU
+            vitaminE: Double.random(in: 0...30), // Random vitamin E between 0 and 30 mg
+            vitaminK: Double.random(in: 0...300), // Random vitamin K between 0 and 300 mcg
             vitaminB1: Double.random(in: 0...5), // Random vitamin B1 between 0 and 5 mg
             vitaminB2: Double.random(in: 0...5), // Random vitamin B2 between 0 and 5 mg
-            vitaminB3: Int.random(in: 0...50), // Random vitamin B3 between 0 and 50 mg
-            vitaminB5: Int.random(in: 0...100), // Random vitamin B5 between 0 and 100 mg
+            vitaminB3: Double.random(in: 0...50), // Random vitamin B3 between 0 and 50 mg
+            vitaminB5: Double.random(in: 0...100), // Random vitamin B5 between 0 and 100 mg
             vitaminB6: Double.random(in: 0...5), // Random vitamin B6 between 0 and 5 mg
-            vitaminB7: Int.random(in: 0...300), // Random vitamin B7 between 0 and 300 mcg
-            vitaminB9: Int.random(in: 0...1000), // Random vitamin B9 between 0 and 1000 mcg
+            vitaminB7: Double.random(in: 0...300), // Random vitamin B7 between 0 and 300 mcg
+            vitaminB9: Double.random(in: 0...1000), // Random vitamin B9 between 0 and 1000 mcg
             vitaminB12: Double.random(in: 0...5) // Random vitamin B12 between 0 and 5 mcg
         )
         return meal
